// src/pages/message-management/enquire-and-manage-message/type.ts
export interface IManageMessage {
  id: string;
  smaMessageId: string;
  cmmpMessageId: string;
  sentDateAndTime: string;
  toS: string;
  ccS: string;
  bccS: string;
  deliveryStatus: string;
  sentDate: string;
  classification: string;
  messageSize: string;
  subject: string;
  sender: string;
  recipient: string;
  relatedMessage: string;
}